feat(driver): Disable malloc from uart16550_recvfile() method.
There are cases where we want to receive files starting at address 0
(for example, to initialize the memory of iob_system). The malloc code
would assume address 0 was NULL, and write to another place instead.

The driver of the iob-uart peripheral (https://github.com/IObundle/iob-uart)
also does not have this malloc code.